% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get_interval_bounds.R
\name{get_interval_bounds}
\alias{get_interval_bounds}
\title{get_interval_bounds}
\usage{
get_interval_bounds(x, name_prefix = "")
}
\arguments{
\item{x}{<character vector> character vector as returned by the "cut" function in BASE R.}

\item{name_prefix}{<string> a prefix to be included for the column names of the output variables (representing the lower limit and the upper limit of the bin). If this is "", the default,
then the output column names are simply "inf" and "sup". If, for example you put name_prefix = "latitude", then the output names will be "latitude_inf" and "latitude_sup".}
}
\value{
<data frame> this returns a data frame with two columns. The present function is meant to be used within a tidyr pipe
(within dplyr::mutate, more specifically). For example:
"mydf = mydf \%>\% dplyr::mutate(get_interval_bounds(latitude, name_prefix = "latitude"))"
}
\description{
This utility function is to be used together with the "cut" function of base R. The "cut" function allows you to "bin" a continuous numerical variable, returning a character vector as output.
Often, instead of a string describing the bin you want instead to have the lower limit and the upper limit for each bin that is returned by the "cut" function. The present function does precisely that.
}
\examples{

test_vec = cut(1:10, seq(0,10,2))
get_interval_bounds(test_vec, "var")

}
