% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_plot_lat_surface_on_map.R
\name{plot_lat_surface_on_map}
\alias{plot_lat_surface_on_map}
\title{plot_lat_surface_on_map}
\usage{
plot_lat_surface_on_map(latselect, xlim = c(-180, 180), ylim = c(-90, 90))
}
\arguments{
\item{latselect}{<numerical vector> latitudes to be plotted on map}

\item{xlim}{<c(lon_min, lon_max)> longitude limits of the map to be plotted}

\item{ylim}{<c(lat_min, lat_max)> latitude limits of the map to be plotted}
}
\value{
a ggplot object
}
\description{
visualizes the latitudes of the selected latitudinal surfaces on a map.
}
\examples{
#none
}
